Fuzzy-set if no exact match
Currently,

```
$ kubens def
error: no namespace exists with name "def".
```

This commit instead drops the user into fzf, where the first result is
probably 'default', which they can hit enter to accept. If it is the
only result, it is *not* accepted by default, giving the user a chance
to confirm.

fzf is only used according to the existing rules: installed, stdout is a
tty, and disabling env var not set.