Hello :)
It's my first PR to an Go project here, so be kind to me :)

It's possible to set KUBECTX_DIR and KUBENS_DIR as an environemnt
variable.

Example:

> KUBECTX_DIR="/foo/bar"

will result in /foo/bar/.kube/kubectx

When KUBECTX_DIR is not set, kubectx and kubens fall back
to the default behavior.

issue: #402